tests/requirements.txt
boto3>=1.34.0
pyyaml>=6.0
pytest>=8.0
